+++ src/components/pages/ProjectDetails/ProjectDetails.tsx
@@ -2,7 +2,7 @@ import axios from 'axios';
 import { useEffect, useState, useContext } from 'react';
 import Swal from 'sweetalert2';
 import { useNavigate, useParams } from 'react-router-dom';
-import api from '../../../api/index';
+import api from '../../../store/api/index';
 import { CartContext } from '../../../context/CartContext/CartContext';
 import { useProject } from '../../../context/ProjectContext';
 import { ISpecies, IProjectTree } from '../../../@types';
